:art: refactor: rename and move `PrecalcSqlQueryBuilder`
